Fix `long double` issue:
In RV64 there are no `long double` registers (if we suppose they are 16
Bytes wide), as registers have only 8 bytes. During the backport I
missed some extra code needed in tccgen.c, because it needs special
handling of the 16 Byte case.

Before this, fix, the following program was failing in the assertion at
riscv64-gen.c:449 (`assert(size == 4 || size == 8)`), as it was
expecting only to use `double`s or `float`s:

```
long double f(int a){
    return  a;
}

int main(int argc, char*argv[]){
    return 1;
}
```

PROBLEM: with the changes in this commit, it fails with:

```
tcc: error: undefined symbol '__floatsitf'
```

That symbol is defined in `lib/lib-arm64.c`, which is reused in RV64.
We need to check it because it's probably loaded incorrectly or not
loaded at all.