RISC-V: Cast ints to 32 bit integers when needed
Mes defaults to 64 bit integers, which is perfectly OK according to the
C standard. TCC supposes ints are 32 bits, which is not necessarily true
and in Mes everything falls apart.

Casting intermediate values avoid this kind of issue and make TCC more
stable throughout C implementations. If we don't do this:

0b1001 << 30 >> 30
^^^^^^^^^^^^
this overflows in 32 bit integers but in 64 bit integers it doesn't.

Casting avoids this issue, truncating intermediate values.