RISCV: bootstrap: Use correct flags to produce a working libtcc1.a.
We were generating good enough libtcc1.a for building next steps in the
bootstrap.sh program but the last one missed symbols. We didn't detect
this because we didn't build anything after it. Until now: tcc-mob
complained about missing symbols.

With this patch we add the necessary flags to the last libtcc1.a
generation, that were missing.

* bootstrap.sh: Add proper build flags in riscv64 rebuilds